Stable demo code v0.4

+++ FiniteRoad.m
@@ -29,8 +29,8 @@ classdef FiniteRoad < Road
                 obj.spawnTimeProbDist = makedist('Exponential','mu',finite_road_args{2});
                 obj.carTypePd = makedist('uniform',0,1);
                 obj.spawn_car(0,finite_road_args{4});
-                obj.numCarsHistory = NaN(finite_road_args{5},1);
-                obj.averageVelocityHistory = NaN(finite_road_args{5},1);                
+                obj.numCarsHistory = NaN(int8(finite_road_args{5}),1);
+                obj.averageVelocityHistory = NaN(int8(finite_road_args{5}),1);
             end
         end
         function spawn_car(obj,time,dt)
